requirements.txt
google-cloud-bigquery==3.25.0
altair==5.3.0
plotly==5.24.1
uuid==1.30.0
Authlib==1.3.2
firebase_admin==6.8.0
xlsxwriter==3.2.3
extra_streamlit_components~=0.1.80
requests==2.32.5
